requirements.txt
django>=4.2.0
openai>=1.9.0
python-dotenv>=1.0.0
whitenoise>=6.5.0
